+++ packages/app/components/map/MapContainer.web.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { View, Platform } from 'react-native';
 
-import WebMap from './WebMap';
+import WebMap from './WebMap.web';
 import { isObjectEmpty } from '../../utils/isObjectEmpty';
 import { defaultShape } from '../../utils/mapFunctions';
 import useCustomStyles from 'app/hooks/useCustomStyles';
