+++ packages/app/screens/trip/useWeather.tsx
@@ -1,0 +1,10 @@
+import { useSelector } from 'react-redux';
+
+const useWeather = () => {
+  const weatherObject = useSelector((state) => state.weather.weatherObject);
+  const weatherWeek = useSelector((state) => state.weather.weatherWeek);
+
+  return { weatherObject, weatherWeek };
+};
+
+export default useWeather;
